move method in rigth file
