+++ packages/code-du-travail-modeles/src/modeles/common/utils/__tests__/accumulate-absence-by-year.spec.ts
